move parsed query type to ir module and rename it to Query
