+++ fluree-db-api/src/format/construct.rs
@@ -12,7 +12,7 @@ use super::{FormatError, Result};
 use crate::QueryResult;
 use fluree_db_core::FlakeValue;
 use fluree_db_query::binding::Binding;
-use fluree_db_query::parse::ConstructTemplate;
+use fluree_db_query::ir::ConstructTemplate;
 use fluree_db_query::ir::triple::{Ref, Term};
 use fluree_db_query::Batch;
 use fluree_graph_format::{format_jsonld, JsonLdFormatConfig};
