+++ fluree-db-api/src/format/sparql_xml.rs
@@ -355,7 +355,7 @@ mod tests {
             novelty: None,
             context: fluree_graph_json_ld::ParsedContext::default(),
             orig_context: None,
-            output: fluree_db_query::parse::QueryOutput::select(vec![]),
+            output: fluree_db_query::ir::QueryOutput::select(vec![]),
             batches: vec![],
             binary_graph: None,
             graph_select: None,
@@ -366,7 +366,7 @@ mod tests {
     fn format_ask_true() {
         let compactor = make_test_compactor();
         let mut result = make_test_result();
-        result.output = fluree_db_query::parse::QueryOutput::Boolean;
+        result.output = fluree_db_query::ir::QueryOutput::Boolean;
         result.batches = vec![Batch::single_empty()];
 
         let xml = format(&result, &compactor, &FormatterConfig::sparql_xml()).unwrap();
@@ -380,7 +380,7 @@ mod tests {
         let mut result = make_test_result();
 
         let s_var = result.vars.get_or_insert("?s");
-        result.output = fluree_db_query::parse::QueryOutput::select(vec![s_var]);
+        result.output = fluree_db_query::ir::QueryOutput::select(vec![s_var]);
 
         let schema = std::sync::Arc::from(vec![s_var].into_boxed_slice());
         let sid = Sid::new(100, "alice");
