+++ fluree-db-api/src/query/mod.rs
@@ -17,7 +17,7 @@ use crate::{
 use fluree_db_binary_index::BinaryGraphView;
 use fluree_db_core::{GraphDbRef, LedgerSnapshot};
 
-use fluree_db_query::parse::{GraphSelectSpec, QueryOutput};
+use fluree_db_query::ir::{GraphSelectSpec, QueryOutput};
 
 /// Result of a query execution
 pub struct QueryResult {
