+++ fluree-db-query/src/ir.rs
@@ -5,16 +5,21 @@
 //!
 //! # Design
 //!
-//! - `Query` is the top-level structure with `select` and ordered `where_` patterns
+//! - `Query` is the top-level structure: output spec, WHERE patterns, options,
+//!   optional graph crawl spec, optional post-VALUES
 //! - `Pattern` enum mirrors the where clause structure, preserving order for filter inlining
 //! - The planner chooses physical join operators based on pattern analysis
 //!
 //! # Module layout
 //!
+//! - [`query`] — top-level `Query`, `QueryOutput`, `ProjectionShape`,
+//!   `ConstructTemplate`: the resolved-and-lowered query that flows through
+//!   parse → plan → execute → format
 //! - [`triple`] — `TriplePattern`, `Ref`, `Term`: the s/p/o vocabulary used
 //!   by triple patterns (and reused by other pattern variants for s/p
 //!   positions)
-//! - [`projection`] — top-level `Query` and graph-crawl selection specs
+//! - [`projection`] — graph-crawl selection specs (`GraphSelectSpec`,
+//!   `NestedSelectSpec`, `SelectionSpec`, `Root`)
 //! - [`path`] — property-path patterns (transitive predicate traversal)
 //! - [`adapters`] — scan patterns over non-graph data sources (BM25, vector,
 //!   geo, S2, R2RML) adapted to plug into the pattern tree
@@ -28,6 +33,7 @@ pub mod expression;
 pub mod path;
 pub mod pattern;
 pub mod projection;
+pub mod query;
 pub mod triple;
 
 pub use adapters::{
@@ -40,4 +46,5 @@ pub use pattern::{
     GraphName, Pattern, ServiceEndpoint, ServicePattern, SubqueryPattern,
 };
 pub use projection::{GraphSelectSpec, NestedSelectSpec, Root, SelectionSpec};
+pub use query::{ConstructTemplate, ProjectionShape, Query, QueryOutput};
 pub use triple::{Ref, Term, TriplePattern};
