+++ fluree-db-query/src/parse/ast.rs
@@ -1010,7 +1010,7 @@ pub struct UnresolvedQuery {
     /// (array form, or SPARQL, which has no scalar form). Propagated into
     /// `QueryOutput::Select` / `SelectOne` so formatters can decide rendering
     /// without inferring from `vars.len()`.
-    pub select_shape: crate::parse::lower::ProjectionShape,
+    pub select_shape: crate::ir::ProjectionShape,
     /// Ordered patterns in where clause (triples, filters, optionals, etc.)
     pub patterns: Vec<UnresolvedPattern>,
     /// Query options (limit, offset, order by, group by, etc.)
@@ -1031,7 +1031,7 @@ impl UnresolvedQuery {
             context,
             orig_context: None,
             select: Vec::new(),
-            select_shape: crate::parse::lower::ProjectionShape::default(),
+            select_shape: crate::ir::ProjectionShape::default(),
             patterns: Vec::new(),
             options: UnresolvedOptions::default(),
             construct_template: None,
